[Refactor] api uri 명칭 변경 및 Building 패키지로 이동
